feat: Operator열거형 선언, String의 split 확장, Formula 기본형태 구축

+++ Calculator/CalculatorTests/CalculatorTests.swift
@@ -10,7 +10,7 @@ import XCTest
 
 final class CalculatorTests: XCTestCase {
 
-    var sut: CalculatorItemQueue<Int>!
+    var sut: CalculatorItemQueue<Double>!
     
     override func setUpWithError() throws {
         try super.setUpWithError()
